+++ pcp_stats.h
@@ -112,7 +112,7 @@ void pcp_register_sadc_metrics(void);
 void pcp_write_sadc_header(long interval_secs);
 void pcp_write_sadc_info_file(const char *archive_path);
 void pcp_write_sadc_special_record(const char *comment, unsigned int cpu_nr,
-				   unsigned long long timestamp, long nsec);
+				   unsigned long long timestamp, unsigned int nsec);
 
 /* Shared PCP archive reading helpers (used by sar and sadf) */
 void check_pcpfile_actlist(const char *from_file, struct activity *act[], uint64_t flags);
@@ -135,7 +135,7 @@ void pcp_write_sadf_comment(const char *comment, unsigned long long ust_time);
 #ifdef HAVE_PMI_APPEND
 int  pcp_open_sadc_archive(const char *path, const struct file_header *hdr);
 void pcp_write_uptime(unsigned long long uptime_cs);
-int  pcp_write_sadc_sample(unsigned long long ust_time, long nsec,
+int  pcp_write_sadc_sample(unsigned long long ust_time, unsigned int nsec,
 			   uint64_t flags);
 void pcp_sadc_set_volume_size(size_t volume_size);
 void pcp_close_sadc_archive(void);
